+++ src/main/kotlin/com/atlassian/performance/tools/infrastructure/api/jvm/Jstat.kt
@@ -1,6 +1,7 @@
 package com.atlassian.performance.tools.infrastructure.api.jvm
 
 import com.atlassian.performance.tools.infrastructure.api.os.MonitoringProcess
+import com.atlassian.performance.tools.infrastructure.api.process.RemoteMonitoringProcess
 import com.atlassian.performance.tools.ssh.api.SshConnection
 import java.time.Duration
 
@@ -19,11 +20,30 @@ class Jstat(
                 "done"
     }
 
+    @Deprecated("Use `start` method.")
     fun startMonitoring(
         connection: SshConnection,
         pid: String,
         option: String = "-gcutil"
     ): MonitoringProcess {
+        val interval = "${INTERVAL.seconds}s"
+        val process = connection.startProcess("${jvmBin}jstat $option -t $pid $interval | $ADD_TIME > $LOG_PATH")
+        return MonitoringProcess(process, LOG_PATH)
+    }
+
+    fun start(
+        connection: SshConnection,
+        pid: Int
+    ): RemoteMonitoringProcess {
+        val option = "-gcutil"
+        return start(connection, pid, option)
+    }
+
+    fun start(
+        connection: SshConnection,
+        pid: Int,
+        option: String
+    ): RemoteMonitoringProcess {
         val interval = "${INTERVAL.seconds}s"
         val process = connection.startProcess("${jvmBin}jstat $option -t $pid $interval | $ADD_TIME > $LOG_PATH")
         return MonitoringProcess(process, LOG_PATH)